+++ lib/src/page/divider_section.dart
@@ -1,4 +1,6 @@
 import 'package:demoflu/src/page/page_section.dart';
+import 'package:demoflu/src/provider.dart';
+import 'package:demoflu/src/theme.dart';
 import 'package:flutter/material.dart';
 
 /// Section to display a divider.
@@ -6,10 +8,20 @@ class DividerSection extends PageSection {
   DividerSection(
       {required super.marginLeft,
       required super.marginBottom,
-      required super.maxWidth});
+      required super.maxWidth,
+      required this.thickness,
+      required this.color});
+
+  Color? color;
+  double? thickness;
 
   @override
   Widget buildContent(BuildContext context) {
-    return Divider(height: 1, color: Colors.grey[300], thickness: 1);
+    DemoFluTheme theme = DemoFluProvider.themeOf(context);
+    double thickness = this.thickness ?? theme.dividerThickness;
+    return Divider(
+        height: thickness,
+        color: color ?? theme.dividerColor,
+        thickness: thickness);
   }
 }