+++ lib/src/page/page.dart
@@ -1,9 +1,6 @@
+import 'package:demoflu/demoflu.dart';
 import 'package:demoflu/src/macro.dart';
-import 'package:demoflu/src/page/borders/section_border.dart';
-import 'package:demoflu/src/page/heading_section.dart';
-import 'package:demoflu/src/page/page_section_group.dart';
 import 'package:demoflu/src/page/section_collection.dart';
-import 'package:demoflu/src/page/text_section.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
@@ -17,22 +14,52 @@ abstract class DemoFluPage with SectionCollectionMixin {
 
   /// Runs a  macro.
   void runMacro(dynamic id, BuildContext context) {
-    _macroFactory.getPageMacro(id)(context, this);
+    _macroFactory.getMacro<PageMacro>(id, 'Page')(context, this);
   }
 
   /// Runs a group section macro.
   PageSectionGroup runGroupMacro(dynamic id, BuildContext context) {
-    return _macroFactory.getGroupMacro(id)(context, this);
+    return _macroFactory.getMacro<GroupMacro>(id, 'Group')(context, this);
   }
 
   /// Runs a heading section macro.
   HeadingSection runHeadingMacro(dynamic id, BuildContext context) {
-    return _macroFactory.getHeadingMacro(id)(context, this);
+    return _macroFactory.getMacro<HeadingMacro>(id, 'Heading')(context, this);
   }
 
   /// Runs a heading section macro.
   TextSection runTextMacro(dynamic id, BuildContext context) {
-    return _macroFactory.getTextMacro(id)(context, this);
+    return _macroFactory.getMacro<TextMacro>(id, 'Text')(context, this);
+  }
+
+  /// Runs a heading section macro.
+  BulletsSection runBulletsMacro(dynamic id, BuildContext context) {
+    return _macroFactory.getMacro<BulletsMacro>(id, 'Bullets')(context, this);
+  }
+
+  /// Runs a divider section macro.
+  DividerSection runDividerMacro(dynamic id, BuildContext context) {
+    return _macroFactory.getMacro<DividerMacro>(id, 'Divider')(context, this);
+  }
+
+  /// Runs a banner section macro.
+  BannerSection runBannerMacro(dynamic id, BuildContext context) {
+    return _macroFactory.getMacro<BannerMacro>(id, 'Banner')(context, this);
+  }
+
+  /// Runs a widget section macro.
+  WidgetSection runWidgetMacro(dynamic id, BuildContext context) {
+    return _macroFactory.getMacro<WidgetMacro>(id, 'Widget')(context, this);
+  }
+
+  /// Runs a code section macro.
+  CodeSection runCodeMacro(dynamic id, BuildContext context) {
+    return _macroFactory.getMacro<CodeMacro>(id, 'Code')(context, this);
+  }
+
+  /// Runs a console section macro.
+  ConsoleSection runConsoleMacro(dynamic id, BuildContext context) {
+    return _macroFactory.getMacro<ConsoleMacro>(id, 'Console')(context, this);
   }
 
   /// Creates a group of sections.